fix(compartment-mapper): dynamic requires via absolute path use last-resort package policy lookup
This adds `enforcePackagePolicyByPath()` which accepts two (2) `CompartmentDescriptor`s (_A_ and _B)_ and determines if _B_ is allowed access to _A_ via package policy.  `findRedirect()`, called by `importNowHook()`, now uses this if it cannot enforce policy any other way.  This solves the issue described in #2876, which is that a package needs to dynamically require a module from an otherwise-unrelated-except-by-policy `CompartmentDescriptor`.

Slightly changed the existing error hint to differentiate this case.

Fixes #2876