fix(compartment-mapper): guarantee stable paths (#2872)
Previously, it was trivially possible to generate a dependency tree which caused `mapNodeModules` to return incorrect shortest values of `CompartmentDescriptor.path`.  An example such tree (which is used in test fixtures):

```mermaid
graph TD
    app --> paperino
    app --> pippo
    paperino --> topolino
    pippo --> gambadilegno
    topolino --> gambadilegno
    gambadilegno --> goofy
```

In the above tree, there are two paths from `app` to `goofy`; `['app', 'pippo', 'gambadilegno', 'goofy']` and `['app', 'paperino', 'topolino', 'gambadilegno', 'goofy']`.  The shortest path from `app` to `goofy` is the former, but `mapNodeModules` would consistently return the latter due to its depth-first nature and real-time calculation of optimal paths.  In addition, it was possible for larger dependency trees to be completely non-deterministic due to asynchronous I/O operations.

I decided the current strategy was too error-prone to fix _in situ_, so I tore it out.  The result is this change, which fixes `@endo/compartment-mapper` so that `mapNodeModules` will return a `CompartmentMapDescriptor` with stable and correct shortest values for `path`.

This was done by implementing a graph data structure which maps nodes (packages) via edges (dependee -> dependency) using an edge weighting function based only on the dependency name (which is analogous to order of comparisons done in the `@endo/path-compare` algorithm [I think]).

When the filesystem crawl is complete—and prior to translation into a `CompartmentMapDescriptor`—we calculate the shortest path from the entry to each other member of the graph (excepting the attenuator node, if it exists), and use this information to set the `path` value for the data structure entering `translateGraph`.