feat(compartment-mapper): support mapping of undiscoverable packages
This adds options to `mapNodeModules()`, `captureFromMap()`, `loadFromMap()` and `importLocation()` to support loading additional packages that wouldn't otherwise be discovered by `mapNodeModules()`.  It takes several ideas from #2864 which were not already implemented in #2876.

The first option is `additionalModuleLocations`, which is an `Array<string | AdditionalModuleLocationObject>` where items can either be fully-qualified `file://` URLs or an object with shape:

```ts
export interface AdditionalModuleLocationObject {
  location: string;
  dev?: boolean;
  conditions?: Set<string>;
}
```

Using the above object, consumers have granular control over certain options when graphing additional modules.  If `dev` or `conditions` are provided ("not `undefined`") as options to `mapNodeModules()` directly, the additional modules will _inherit_ those settings (there is coverage for this).

The second option, `additionalPackageDetails`, is a superset of `AdditionalModuleLocationObject` which also contains a package descriptor and package location.  This array can then be handed to `captureFromMap()` or `loadFromMap()` after it has been populated.

Because the `CompartmentMapDescriptor` can only have a single `CompartmentDescriptor` with an empty `path` (the entry), the `path` of each additional package is its name.  Related `CompartmentDescriptors`

In `mapNodeModules()`, `additionalModuleLocations` causes `graphPackages()` to be called against each location while reusing the `Graph`. The internal API has been changed somewhat to allow this. Because we do not know whether or we will be executing using the resulting `CompartmentMapDescriptor` (e.g. we can provide it to `captureFromMap()` which will not invoke dynamic requires), `mapNodeModules()` sets the `CompartmentDescriptor.retained` flag. When it does, it also sets the `retained` flag for each `ModuleDescriptor` found within `CompartmentDescriptor.modules`.  It does this as the _last_ step before returning the `CompartmentMapDescriptor` and only does so on `CompartmentDescriptor`s not already found prior to graphing additional modules. This is a recursive operation which bails out of if either a) the `CompartmentDescriptor` was graphed by way of the entry module, or b) we've seen the `CompartmentDescriptor` before (to protect against cycles).

The supporting implementation in other modules is minimal excepting a change to `loadFromMap` where all additional packages are eagerly loaded via `compartment.load(additionalPackageLocation)`.

Tests are concentrated in a new test suite (`additional-modules.test.js`) which covers the aforementioned APIs using the same comprehensive fixture.

Also:

- Fixed signature of `chooseModuleDescriptor`
- Fixed deprecation notice for `compartmentMapForNodeModules`; it was previously on the options object, but it should only be on the export.
- Passed `log` option thru `importLocations`
- Added explicit type for `LoadFromMapOptions` (it was previously using `ImportLocationOptions`, but it is now slightly different)
- `makeModuleMapHook()` now accepts a `Record<string, CompartmentDescriptor>` parameter (ostensibly from `CompartmentMapDescriptor.compartments`) for the purpose of error reporting, which can now reference canonical names / paths