refactor(pass-style): refactor-only avoid symbol-named methods (#2840)
Extracted from #2797 

Closes: #XXXX
Refs: #2793 #2792

## Description

This PR is the extraction from #2797 of only the safe pure refactor
elements of #2797 as suggested by @kriskowal at
https://github.com/endojs/endo/pull/2797#pullrequestreview-2865819668 .
#2797 is then rebased on this one, representing only the elements whose
refactor-purity was questionable.

This PR should be a pure refactor, to avoid creating remotables with
symbol-named methods when there is clearly no possible compat hazard.
Non-remotable objects with symbol-named methods are fine. But see
Compatibility Considerations below.

### Security Considerations

none
### Scaling Considerations

none
### Documentation Considerations

none
### Testing Considerations

Most of the remotable objects with symbol-named methods were in tests.
These can clearly be fixed without any compat hazard, so I count these
as pure refactors. However, in removing the symbol-ness of some of these
tests, I may have accidentally removed the purposes of the tests. They
might now be fully redundant with other tests and should be removed. If
anyone does investigate this, that cleanup could be in a later PR.

### Compatibility Considerations

none. But see Compatibility Considerations of #2797 

### Upgrade Considerations

none