fix(pass-style): revert precision of deprecated isObject type (#2892)
Closes: #XXXX
Refs: https://github.com/endojs/endo/pull/2860

## Description

#2860 deprecated `isObject` as confusing and promoted use of
`isPrimitive` instead. However, it also made the type of `isObject` more
precise, which accidentally broken a bunch of client code in agoric-sdk
(and possibly elsewhere) that relied on the type being less precise.
Since `isObject` is now deprecated and all clients should eventually
switch to `!isPrimitive` anyway, it costs nothing to lose this extra
type precision of `isObject`.

This PR restores the old imprecise type.

### Security Considerations

none
### Scaling Considerations

none
### Documentation Considerations

none
### Testing Considerations

none
### Compatibility Considerations

The point. This PR restores the compatibility that #2860 accidentally
broke.
### Upgrade Considerations

none