feat(ses): repair missing "hasOwn" (#2927)
Closes: https://github.com/endojs/endo/issues/2930
Refs: #XXXX

## Description

`Object.hasOwn` is supported on all JS engine versions that Endo
officially supports, such as Node 18. Thus, technically, Endo should not
need to cope with the absence of `Object.hasOwn`. However, as
https://github.com/endojs/endo/issues/2930 reports, this absence seems
to be the only impediment to running on the version of Safari/JSC
shipping on iOS 15.3 or earlier. Fortunately, `Object.hasOwn` is trivial
to shim, so this PR trivially shims it when it is absent.

### Security Considerations

Supporting such ancient engines makes it more likely for people to run
ses on such ancient engines, which is indeed the point of this PR.
However, ancient engines are also more likely to violate conformance
with modern JS in other ways that may violate the ses implementation's
assumptions, which is a security hazard.

### Scaling Considerations

practically none
### Documentation Considerations

none
### Testing Considerations

Rather than arrange to test the Safari/JSC from iOS 15.3 under CI, the
test cases added by this PR emulates those conditions by deleting
`Object.hasOwn` before `commons.js` initializes.

### Compatibility Considerations

The point. The restores the compat of Endo with platforms lacking their
own `Object.hasOwn`, assuming that was their only incompat with modern
ses.

### Upgrade Considerations

None

- [ ] Update `NEWS.md` for user-facing changes.