fix(ses): to workaround #2908 use current console (#2934)
Closes: #2908
Refs: #636 #944 #945 https://github.com/nodejs/node/pull/59456

## Description

Due to timing, reporter.js runs before lockdown replaces the system
console with our causal console. Thus, when `makeReporter` makes a
reporter that uses some global console, it uses the platform's original
global console rather than the causal console. I don't know if this was
purposeful or good enough at the time, so I don't know whether this PR
is a workaround or a proper fix. In any case, it fixes the case of
immediate interest by using the current global console to report rather
than the original one.

### Security Considerations

@gibson042 checked below and found that the `console` property of the
global of the start compartments remains mutable after lockdown. This is
a security hazard, but is consistent with our stance on the privileges
available in the start compartment. In any case, this hazard is not
affected by this PR.

### Scaling Considerations

none
### Documentation Considerations

none
### Testing Considerations

Hard to write a regression test to ensure that a stack is output since
there is no portable agreement on what stacks look like. But not
impossible.

Manually running the test case from #2908 , we get the correct outputs
at https://github.com/endojs/endo/issues/2908#issuecomment-3180527514 .

Running that test case from #2908 on Node 22 and 24, we find that the
bug does not manifest there either. The problem is ***only*** on Node
20. I don't understand how that can be. In any case, this PR does fix
the behavior across all these versions.

### Compatibility Considerations

If someone suppresses the `console` replacement with `consoleTaming:
'unsafe'`, then this PR in its current state fails to workaround #2908

If some code depends on reporters using the original `console` rather
than the current one, that code might break or misbehave.

### Upgrade Considerations

none