feat(errors): confirm/reject pattern (as suggested by #2285) (#2923)
Closes: #2285
Refs: #2851 #2917 #2918 #2917 #2922

## Description

Follows approximately @gibson042 's suggestions at #2285, with some
minor differences I'd like @gibson042 's feedback on:

- I changed the template type in the `Rejector` type to end with `=>
never` rather than `=> false`.
- I added a `{false}` type disjunct to the `Rejector` type
- made the `reject` parameter to the `confirmFoo` functions
non-optional, since it would also be supplied by the `isFoo` or
`assertFoo` functions. Typically, only the latter two would be exported.
The only reason I can see to make `confirmFoo`'s `reject` parameter
optional is to be able to export and use it directly as the predicate,
which would be an abstraction leak.

As stated in #2285, the purpose is to migrate existing uses of `Checker`
to use `Rejector` instead. The first real tests for that is #2917 to use
Rejectors instead of Checkers in @endo/pass-style, and #2922 to do
likewise for @endo/patterns.

### Security Considerations

Since this PR itself is only additive, if it is correct, none. If so,
then #2917 and #2922 hopefully should improve security since code using
Rejectors should overall be simpler and easier to review for correctness
that code using Checkers.

### Scaling Considerations

Most of the point. We hope that code using Rejectors will be faster than
current code using Checkers. We should measure the speedups of #2917 and
#2922 over `master` baseline.

### Documentation Considerations

Most code uses only `isFoo` or `assertFoo` functions, whose behavior
should be identical. Only users or definers of the old `checkFoo` or the
new `confirmFoo` functions need to understand the difference. Not sure
where these are or should be documented.

### Testing Considerations

The test included in this PR also demonstrate the
`confirm/reject/is/assert` pattern that Rejectors are designed to
support.

### Compatibility Considerations

For this PR by itself, since it is only additive, none. Further, the
design of Rejectors is intended to enable code to switch over without
causing external compat considerations. The redo of #2851 will be a good
first test.

For any old `checkFoo` code using Checkers rather than Rejectors, if
those `checkFoo` functions were exported, we can deprecate and continue
to export them during the transition to further avoid compat issues.

### Upgrade Considerations

None of this causes any change to data that is communicated or stored.
Given that the `isFoo` and `assertFoo` functions are behaviorally
identical and that any exported `checkFoo` functions continue to be
exported (as deprecated), there should be no upgrade concerns.