refactor(pass-style): use confirm/reject pattern in pass-style (#2917)
Closes: #XXXX
Refs: #2911 #2285 #2918 #2922 

## Description

Uses the new Rejector pattern suggested by #2285 and implemented in
#2923 to restructure all the check*/Checker patterns in @endo/pass-style
to use the confirm*/reject pattern instead. This is also a test and
demonstration of the basic functionality of #2923.

For this experiment, as @gibson042 expected, the code is both smaller
and more readable.

Should be a pure refactor, not causing any effects observable from
outside @endo/pass-style

### Security Considerations

If it is indeed a pure refactor, none. Except for course that the code
is simpler, smaller, and easier to understand and review, which helps
security.

### Scaling Considerations

From eyeballing these, I suspect `passStyleOf` will be faster, but need
to measure that to see how much. This could be significant, as previous
measurement have consistently shown `passStyleOf` to be a bottleneck.

### Documentation Considerations

The actual confim/reject pattern needs to be documented somewhere, for
developers that wish to use it internally. But since the confirm*
functions/methods normally encapsulated, as they are here, normal
developers who consume our APIs won't need to be aware of these changes.

### Testing Considerations

No tests needed to change, raising our confidence that it is a pure
refactor.

### Compatibility Considerations

As a pure refactor with no change to the representation of any data,
should be none.

There remains the linking problem if these changes to @endo/patterns are
linked with a version of @endo/errors prior to #2923. This PR does pull
in the `Rejector` type exported by @endo/errors/rejector.js, which does
not exist in earlier forms of @endo/errors. But these are only static
types and static type imports, so it is hard to see how it could cause a
dynamic problem.

### Upgrade Considerations

As a pure refactor, none.