fix(errors): type `hideAndHardenFunction` like `harden` (#2964)
Closes: #XXXX
Refs: #XXXX

## Description

`hideAndHardenFunction` was intended to be a drop in replacement of
`harden` for functions that you want to hide from compact stacktraces.
It is dynamically. But it did not have the same type as `harden` which
we discovered when such a replacement caused a static type error. This
PR fixes that.

### Security Considerations

static-only change. None

### Scaling Considerations

static-only change. None

### Documentation Considerations

no new doc considerations.

### Testing Considerations

Encouraging more use of `hideAndHardenFunction` will cause concise stack
traces to be less noisy, helping the test and debug cycle.

### Compatibility Considerations

The point. Fixes the one incompat between `harden` and
`hideAndHardenFunction`

### Upgrade Considerations

none