chore: undo changes to archive subsystem
This reverts changes I made where the archive compartment map descriptors were keyed on canonical name. Now they are once again keyed on `packageName-v<version>` or `packageName-v<version>-n#`.

Given that this change was actually in the digest functionality, I had to poke a hole in it so that the caller can choose how to rename the compartment descriptors in the compartment map.  The default behavior is to use the canonical name as the new name, but `archive-lite.js` has an implementation of a `CompartmentsRenameFn` which retains the old behavior.  Because this special string is no longer computed within `mapNodeModules`, I added a `version` property to `PackageCompartmentDescriptor`, which comes out of the package descriptor when we translate the graph into a `PackageCompartmentMap`.  The custom `CompartmentsRenameFn` uses this property to compute its old "label".

Alternatives considered included actually computing the old label value in `mapNodeModules()`, but I was not sure what to call such a property and `version` was more obvious.  Either way, nothing would use this field except the custom `CompartmentsRenameFn` in `makeArchiveCompartmentMap()`.

Another alternative is to make the `label` field actually configurable instead of always a `CanonicalName`, but that means we would need to revert the removal of the `path` property from `PackageCompartmentDescriptor`, since the code to compute the "old label" needs it for stable sorting purposes.

And yet another would be to "fork" `digestCompartmentMap` into two separate functions; one for archives and one for everything else. We would still need the separate implementations of `CompartmentsRenameFn` _somewhere_, and I do not want to expand the public API surface further for this.